Add support for generation of JSON Compilation Database
New option -compdb instructs FBuild to generate JSON Compilation
Database for the targets specified on the command line instead of
building them.

Generation is performed by recursively traversing all dependencies of
the specified targets and:
* Building all DirectoryList nodes to generate lists of input files.
* Emitting one entry for each input file that is included in ObjectList
and Library nodes. When Unity node is used as an input, DB entries are
generated for all its input files instead of generated unity files.

Generated DB uses newer format (split arguments in the "arguments" array
instead of raw command line in the "command" string) to avoid issues
with different rules for argument quoting and escaping in Clang and
FASTBuild.