Fix assertion when FileIO::GetFiles functions are called on non-existent directory
After symlink-related changes in e12a8d2a GetFiles functions started to
fail with assertion when called on non-existent directory. But calling
them on non-existent directory should be allowed (at least from looking
at the older version of code).
To fix that don't expect lstat call on original pathCopy to succeed.

This fixes TestBuildFBuild and TestDistributed tests in the case when
../tmp/Test directory is missing.