Fix alloc/dealloc mismatch in ToolManifest
This was found by AddressSanitizer.
In two places a FileStream object is allocated with new (obviously) and
deallocated with ::Free by AutoPtr. This case of alloc/dealloc mismatch
is actually bad because FileStream has a non-trivial destructor and in
ToolManifest::Deserialize it could be deleted while file is open which
leads to a handle leak.

To fix this I added a new deleter class that uses FDELETE.
Also DefaultDeletor is renamed to FreeDeletor to expose the fact that it
uses free instead of delete (which people might expect given that the
name is similar to std::default_delete).